Include 2P-ECDSA functionality and create mock electrum server
