+++ client/src/mocks/mod.rs
@@ -1,0 +1,1 @@
+pub mod mock_electrum;
