+++ client/src/wallet/shared_wallet.rs
@@ -1,0 +1,583 @@
+//! Shared Wallet
+//!
+//! Contains key shares of co-owned keys between user and server.
+//! Functionality includes key gen, server back-up and co-signing
+
+use bitcoin::Address;
+use bitcoin;
+use bitcoin::consensus::encode::serialize;
+use bitcoin::network::constants::Network;
+use bitcoin::util::bip143::SighashComponents;
+use bitcoin::{TxIn, TxOut};
+use bitcoin::hashes::{sha256d, hex::FromHex};
+use bitcoin::secp256k1::Signature;
+use curv::elliptic::curves::traits::ECPoint;
+use curv::{BigInt, GE};
+use electrumx_client::{electrumx_client::ElectrumxClient, interface::Electrumx};
+use kms::ecdsa::two_party::MasterKey2;
+use kms::ecdsa::two_party::*;
+use serde_json;
+use std::fs;
+use uuid::Uuid;
+
+use centipede::juggling::proof_system::{Helgamalsegmented, Proof};
+use centipede::juggling::segmentation::Msegmentation;
+use kms::chain_code::two_party::party2::ChainCode2;
+
+use super::super::ecdsa;
+use super::super::ecdsa::types::PrivateShare;
+use super::super::escrow;
+use super::super::utilities::requests;
+use super::super::ClientShim;
+use curv::arithmetic::traits::Converter;
+use hex;
+use itertools::Itertools;
+use std::collections::HashMap;
+use std::str::FromStr;
+
+// TODO: move that to a config file and double check electrum server addresses
+const ELECTRUM_HOST: &str = "ec2-34-219-15-143.us-west-2.compute.amazonaws.com:60001";
+//const ELECTRUM_HOST: &str = "testnetnode.arihanc.com:51001";
+const WALLET_FILENAME: &str = "wallet/wallet.data";
+const BACKUP_FILENAME: &str = "wallet/backup.data";
+
+#[derive(Serialize, Deserialize)]
+pub struct SignSecondMsgRequest {
+    pub message: BigInt,
+    pub party_two_sign_message: party2::SignMessage,
+    pub pos_child_key: u32,
+}
+
+#[derive(Debug, Deserialize, Clone)]
+pub struct GetBalanceResponse {
+    pub address: String,
+    pub confirmed: u64,
+    pub unconfirmed: u64,
+}
+
+#[derive(Debug, Deserialize, Clone)]
+pub struct GetListUnspentResponse {
+    pub height: usize,
+    pub tx_hash: String,
+    pub tx_pos: usize,
+    pub value: usize,
+    pub address: String,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct GetWalletBalanceResponse {
+    pub confirmed: u64,
+    pub unconfirmed: u64,
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct AddressDerivation {
+    pub pos: u32,
+    pub mk: MasterKey2,
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct SharedWallet {
+    pub id: String,
+    pub network: String,
+    pub private_share: PrivateShare,
+    pub last_derived_pos: u32,
+    pub addresses_derivation_map: HashMap<String, AddressDerivation>,
+}
+
+impl SharedWallet {
+    pub fn new(client_shim: &ClientShim, net: &String) -> SharedWallet {
+        let id = Uuid::new_v4().to_string();
+        let private_share = ecdsa::get_master_key(client_shim);
+        let last_derived_pos = 0;
+        let addresses_derivation_map = HashMap::new();
+        let network = net.clone();
+
+        SharedWallet {
+            id,
+            network,
+            private_share,
+            last_derived_pos,
+            addresses_derivation_map,
+        }
+    }
+
+    pub fn rotate(self, client_shim: &ClientShim) -> Self {
+        ecdsa::rotate_master_key(self, client_shim)
+    }
+
+    pub fn backup(&self, escrow_service: escrow::Escrow) {
+        let g: GE = ECPoint::generator();
+        let y = escrow_service.get_public_key();
+        let (segments, encryptions) = self.private_share.master_key.private.to_encrypted_segment(
+            &escrow::SEGMENT_SIZE,
+            escrow::NUM_SEGMENTS,
+            &y,
+            &g,
+        );
+
+        let proof = Proof::prove(&segments, &encryptions, &g, &y, &escrow::SEGMENT_SIZE);
+
+        let client_backup_json = serde_json::to_string(&(
+            encryptions,
+            proof,
+            self.private_share.master_key.public.clone(),
+            self.private_share.master_key.chain_code.clone(),
+            self.private_share.id.clone(),
+        ))
+        .unwrap();
+
+        fs::write(BACKUP_FILENAME, client_backup_json).expect("Unable to save client backup!");
+
+        debug!("(wallet id: {}) Backup wallet with escrow", self.id);
+    }
+
+    pub fn verify_backup(&self, escrow_service: escrow::Escrow) {
+        let g: GE = ECPoint::generator();
+        let y = escrow_service.get_public_key();
+
+        let data = fs::read_to_string(BACKUP_FILENAME).expect("Unable to load client backup!");
+        let (encryptions, proof, client_public, _, _): (
+            Helgamalsegmented,
+            Proof,
+            Party2Public,
+            ChainCode2,
+            String,
+        ) = serde_json::from_str(&data).unwrap();
+        let verify = proof.verify(
+            &encryptions,
+            &g,
+            &y,
+            &client_public.p2,
+            &escrow::SEGMENT_SIZE,
+        );
+        match verify {
+            Ok(_x) => println!("backup verified 🍻"),
+            Err(_e) => println!("Backup was not verified correctly 😲"),
+        }
+    }
+
+    pub fn recover_and_save_share(
+        escrow_service: escrow::Escrow,
+        net: &String,
+        client_shim: &ClientShim,
+    ) -> SharedWallet {
+        let g: GE = ECPoint::generator();
+        let y_priv = escrow_service.get_private_key();
+
+        let data = fs::read_to_string(BACKUP_FILENAME).expect("Unable to load client backup!");
+
+        let (encryptions, _proof, public_data, chain_code2, key_id): (
+            Helgamalsegmented,
+            Proof,
+            Party2Public,
+            BigInt,
+            String,
+        ) = serde_json::from_str(&data).unwrap();
+
+        let sk = Msegmentation::decrypt(&encryptions, &g, &y_priv, &escrow::SEGMENT_SIZE);
+
+        let client_master_key_recovered =
+            MasterKey2::recover_master_key(sk.unwrap(), public_data, chain_code2);
+        let pos_old: u32 = requests::post(client_shim, &format!("ecdsa/{}/recover", key_id)).unwrap();
+
+        let pos_old = if pos_old < 10 { 10 } else { pos_old };
+        //TODO: temporary, server will keep updated pos, to do so we need to send update to server for every get_new_address
+
+        let id = Uuid::new_v4().to_string();
+        let addresses_derivation_map = HashMap::new(); //TODO: add a fucntion to recreate
+        let network = net.clone();
+
+        let new_wallet = SharedWallet {
+            id,
+            network,
+            private_share: PrivateShare {
+                master_key: client_master_key_recovered,
+                id: key_id,
+            },
+            last_derived_pos: pos_old,
+            addresses_derivation_map,
+        };
+
+        new_wallet.save();
+        println!("Recovery Completed Successfully ❤️");
+
+        new_wallet
+    }
+
+    pub fn save_to(&self, filepath: &str) {
+        let wallet_json = serde_json::to_string(self).unwrap();
+
+        fs::write(filepath, wallet_json).expect("Unable to save wallet!");
+
+        debug!("(wallet id: {}) Saved wallet to disk", self.id);
+    }
+
+    pub fn save(&self) {
+        self.save_to(WALLET_FILENAME)
+    }
+
+    pub fn load_from(filepath: &str) -> SharedWallet {
+        let data = fs::read_to_string(filepath).expect("Unable to load wallet!");
+
+        let wallet: SharedWallet = serde_json::from_str(&data).unwrap();
+
+        debug!("(wallet id: {}) Loaded wallet to memory", wallet.id);
+
+        wallet
+    }
+
+    pub fn load() -> SharedWallet {
+        SharedWallet::load_from(WALLET_FILENAME)
+    }
+
+    pub fn send(
+        &mut self,
+        to_address: String,
+        amount_btc: f32,
+        client_shim: &ClientShim,
+    ) -> String {
+        let selected = self.select_tx_in(amount_btc);
+        if selected.is_empty() {
+            panic!("Not enough fund");
+        }
+
+        let to_btc_adress = bitcoin::Address::from_str(&to_address).unwrap();
+
+        let txs_in: Vec<TxIn> = selected
+            .clone()
+            .into_iter()
+            .map(|s| bitcoin::TxIn {
+                previous_output: bitcoin::OutPoint {
+                    txid: sha256d::Hash::from_hex(&s.tx_hash).unwrap(),
+                    vout: s.tx_pos as u32,
+                },
+                script_sig: bitcoin::Script::default(),
+                sequence: 0xFFFFFFFF,
+                witness: Vec::default(),
+            })
+            .collect();
+
+        let fees = 10_000;
+
+        let amount_satoshi = (amount_btc * 100_000_000 as f32) as u64;
+
+        let change_address = self.get_new_bitcoin_address();
+
+        let total_selected = selected
+            .clone()
+            .into_iter()
+            .fold(0, |sum, val| sum + val.value) as u64;
+
+        let txs_out = vec![
+            TxOut {
+                value: amount_satoshi,
+                script_pubkey: to_btc_adress.script_pubkey(),
+            },
+            TxOut {
+                value: total_selected - amount_satoshi - fees,
+                script_pubkey: change_address.script_pubkey(),
+            },
+        ];
+
+        let transaction = bitcoin::Transaction {
+            version: 0,
+            lock_time: 0,
+            input: txs_in,
+            output: txs_out,
+        };
+
+        let mut signed_transaction = transaction.clone();
+
+        for i in 0..transaction.input.len() {
+            let address_derivation = self
+                .addresses_derivation_map
+                .get(&selected[i].address)
+                .unwrap();
+
+            let mk = &address_derivation.mk;
+            let pk = mk.public.q.get_element();
+
+            let comp = SighashComponents::new(&transaction);
+            let sig_hash = comp.sighash_all(
+                &transaction.input[i],
+                &bitcoin::Address::p2pkh(
+                    &to_bitcoin_public_key(pk),
+                    self.get_bitcoin_network()).script_pubkey(),
+                (selected[i].value as u32).into(),
+            );
+
+            let signature = ecdsa::sign(
+                client_shim,
+                BigInt::from_hex(&hex::encode(&sig_hash[..])),
+                &mk,
+                BigInt::from(0),
+                BigInt::from(address_derivation.pos),
+                &self.private_share.id,
+            ).unwrap();
+
+            let mut v = BigInt::to_vec(&signature.r);
+            v.extend(BigInt::to_vec(&signature.s));
+
+            let mut sig_vec = Signature::from_compact(&v[..])
+                .unwrap()
+                .serialize_der()
+                .to_vec();
+            sig_vec.push(01);
+
+            let pk_vec = pk.serialize().to_vec();
+
+            signed_transaction.input[i].witness = vec![sig_vec, pk_vec];
+        }
+
+        let mut electrum = ElectrumxClient::new(ELECTRUM_HOST).unwrap();
+
+        let raw_tx_hex = hex::encode(serialize(&signed_transaction));
+        let txid = electrum.broadcast_transaction(raw_tx_hex.clone());
+
+        txid.unwrap()
+    }
+
+    /// sign tx input with key corresponding to given addr
+    pub fn sign_tx_input(&mut self, client_shim: &ClientShim, tx: &mut bitcoin::Transaction, input_index: usize, addr: &Address, amount: u64) {
+        let address_derivation;
+        match self.addresses_derivation_map.get(&addr.to_string()) {
+            None => panic!("No key found in wallet for address"),
+            Some(derivation) => {
+                address_derivation = derivation;
+            },
+        }
+
+        let mk = &address_derivation.mk;
+        let pk = mk.public.q.get_element();
+
+        let comp = SighashComponents::new(&tx);
+        let sig_hash = comp.sighash_all(
+            &tx.input[input_index],
+            &bitcoin::Address::p2pkh(
+                &to_bitcoin_public_key(pk),
+                self.get_bitcoin_network()).script_pubkey(),
+            amount,
+        );
+
+        let signature = ecdsa::sign(
+            client_shim,
+            BigInt::from_hex(&hex::encode(&sig_hash[..])),
+            &mk,
+            BigInt::from(0),
+            BigInt::from(address_derivation.pos),
+            &self.private_share.id,
+        ).unwrap();
+
+        let mut v = BigInt::to_vec(&signature.r);
+        v.extend(BigInt::to_vec(&signature.s));
+
+        let mut sig_vec = Signature::from_compact(&v[..])
+            .unwrap()
+            .serialize_der()
+            .to_vec();
+        sig_vec.push(01);
+
+        let pk_vec = pk.serialize().to_vec();
+
+        tx.input[0].witness = vec![sig_vec, pk_vec];
+
+    }
+
+    pub fn get_new_bitcoin_address(&mut self) -> bitcoin::Address {
+        let (pos, mk) = Self::derive_new_key(&self.private_share, self.last_derived_pos);
+        let pk = mk.public.q.get_element();
+        let address = bitcoin::Address::p2wpkh(
+            &to_bitcoin_public_key(pk),
+            self.get_bitcoin_network()
+        );
+
+        self.addresses_derivation_map
+            .insert(address.to_string(), AddressDerivation { mk, pos });
+
+        self.last_derived_pos = pos;
+
+        address
+    }
+
+    pub fn derived(&mut self) {
+        for i in 0..self.last_derived_pos {
+            let (pos, mk) = Self::derive_new_key(&self.private_share, i);
+
+            let address =
+                bitcoin::Address::p2wpkh(
+                    &to_bitcoin_public_key(mk.public.q.get_element()),
+                    self.get_bitcoin_network()
+                );
+
+            self.addresses_derivation_map
+                .insert(address.to_string(), AddressDerivation { mk, pos });
+        }
+    }
+
+    pub fn get_balance(&mut self) -> GetWalletBalanceResponse {
+        let mut aggregated_balance = GetWalletBalanceResponse {
+            confirmed: 0,
+            unconfirmed: 0,
+        };
+
+        for b in self.get_all_addresses_balance() {
+            aggregated_balance.unconfirmed += b.unconfirmed;
+            aggregated_balance.confirmed += b.confirmed;
+        }
+
+        aggregated_balance
+    }
+
+    // TODO: handle fees
+    pub fn select_tx_in(&self, amount_btc: f32) -> Vec<GetListUnspentResponse> {
+        // greedy selection
+        let list_unspent: Vec<GetListUnspentResponse> = self
+            .get_all_addresses_balance()
+            .into_iter()
+//            .filter(|b| b.confirmed > 0)
+            .map(|a| self.list_unspent_for_addresss(a.address.to_string()))
+            .flatten()
+            .sorted_by(|a, b| a.value.partial_cmp(&b.value).unwrap())
+            .into_iter()
+            .collect();
+
+        let mut remaining: i64 = amount_btc as i64 * 100_000_000;
+        let mut selected: Vec<GetListUnspentResponse> = Vec::new();
+
+        for unspent in list_unspent {
+            selected.push(unspent.clone());
+            remaining -= unspent.value as i64;
+
+            if remaining < 0 {
+                break;
+            }
+        }
+
+        selected
+    }
+
+    pub fn list_unspent(&self) -> Vec<GetListUnspentResponse> {
+        let response: Vec<GetListUnspentResponse> = self
+            .get_all_addresses()
+            .into_iter()
+            .map(|a| self.list_unspent_for_addresss(a.to_string()))
+            .flatten()
+            .collect();
+
+        response
+    }
+
+    /* PRIVATE */
+    fn list_unspent_for_addresss(&self, address: String) -> Vec<GetListUnspentResponse> {
+        let mut client = ElectrumxClient::new(ELECTRUM_HOST).unwrap();
+
+        let resp = client.get_list_unspent(&address).unwrap();
+
+        resp.into_iter()
+            .map(|u| GetListUnspentResponse {
+                value: u.value,
+                height: u.height,
+                tx_hash: u.tx_hash,
+                tx_pos: u.tx_pos,
+                address: address.clone(),
+            })
+            .collect()
+    }
+
+    fn get_address_balance(address: &bitcoin::Address) -> GetBalanceResponse {
+        let mut client = ElectrumxClient::new(ELECTRUM_HOST).unwrap();
+
+        let resp = client.get_balance(&address.to_string()).unwrap();
+
+        GetBalanceResponse {
+            confirmed: resp.confirmed,
+            unconfirmed: resp.unconfirmed,
+            address: address.to_string(),
+        }
+    }
+
+    fn get_all_addresses_balance(&self) -> Vec<GetBalanceResponse> {
+        let response: Vec<GetBalanceResponse> = self
+            .get_all_addresses()
+            .into_iter()
+            .map(|a| Self::get_address_balance(&a))
+            .collect();
+
+        response
+    }
+
+    pub fn get_all_addresses(&self) -> Vec<bitcoin::Address> {
+        let init = 0;
+        let last_pos = self.last_derived_pos;
+
+        let mut response: Vec<bitcoin::Address> = Vec::new();
+
+        for n in init..=last_pos {
+            let mk = self
+                .private_share
+                .master_key
+                .get_child(vec![BigInt::from(0), BigInt::from(n)]);
+            let bitcoin_address = Self::to_p2wpkh_address(&mk, self.get_bitcoin_network());
+
+            response.push(bitcoin_address);
+        }
+
+        response
+    }
+
+    fn derive_new_key(private_share: &PrivateShare, pos: u32) -> (u32, MasterKey2) {
+        let last_pos: u32 = pos + 1;
+
+        let last_child_master_key = private_share
+            .master_key
+            .get_child(vec![BigInt::from(0), BigInt::from(last_pos)]);
+
+        (last_pos, last_child_master_key)
+    }
+
+    fn get_bitcoin_network(&self) -> Network {
+        self.network.parse::<Network>().unwrap()
+    }
+
+    pub fn to_p2wpkh_address(mk: &MasterKey2, network: Network) -> bitcoin::Address {
+        bitcoin::Address::p2wpkh(
+            &to_bitcoin_public_key(mk.public.q.get_element()),
+            network
+        )
+    }
+}
+
+// type conversion
+pub fn to_bitcoin_public_key(pk: curv::PK) -> bitcoin::util::key::PublicKey {
+    bitcoin::util::key::PublicKey {
+        compressed: true,
+        key: pk
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    const TEST_WALLET_FILENAME: &str = "test-assets/wallet.data";
+
+    #[test]
+    fn load_wallet_test() {
+        SharedWallet::load_from(TEST_WALLET_FILENAME);
+    }
+
+    #[test]
+    fn get_address_test() {
+        let  mut w : SharedWallet = SharedWallet::load_from(TEST_WALLET_FILENAME);
+        let a = w.get_new_bitcoin_address();
+        assert!(!a.to_string().is_empty())
+    }
+
+    #[test]
+    fn get_balance_test() {
+        let  mut w : SharedWallet = SharedWallet::load_from(TEST_WALLET_FILENAME);
+        let b = w.get_balance();
+        assert!(b.confirmed > 0);
+    }
+}
